+++ beancount2ledger/hledger.py
@@ -62,12 +62,14 @@ class HLedgerPrinter(LedgerPrinter):
         price_str = ('@ {}'.format(posting.price.to_string(self.dformat_max))
                      if posting.price is not None and posting.cost is None else
                      '')
-
-        # Width we have available for the amount: take width of
-        # flag_posting add 2 for the indentation of postings and
-        # add 2 to separate account from amount
-        len_amount = max(0, 76 - (len(flag_posting) + 2 + 2))
-        posting_str = f'  {flag_posting}  {quote_currency(pos_str):>{len_amount}} {quote_currency(price_str)}'
+        if posting.meta and '__automatic__' in posting.meta and not '__residual__' in posting.meta:
+            posting_str = f'  {flag_posting}'
+        else:
+            # Width we have available for the amount: take width of
+            # flag_posting add 2 for the indentation of postings and
+            # add 2 to separate account from amount
+            len_amount = max(0, 76 - (len(flag_posting) + 2 + 2))
+            posting_str = f'  {flag_posting}  {quote_currency(pos_str):>{len_amount}} {quote_currency(price_str)}'
         self.io.write(posting_str.rstrip())
 
         self.io.write('\n')