+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.2 (unreleased)
 
+* Quote commodities containing minus/dash (`-`) ([issue #15](https://github.com/beancount/beancount2ledger/issues/15))
 * Fix crash with very long account names ([issue #19](https://github.com/beancount/beancount2ledger/issues/19))
 
 ## 1.1 (2020-07-24)